refactor(dir): change directory structure
