+++ pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "iris-software"
+name = "IrisSoftware"
 version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
 readme = "README.md"
-packages = [{include = "iris_software"}]
+packages = [{include = "IrisSoftware"}]
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.11"
